+++ packages/forms/signals/public_api.ts
@@ -20,5 +20,5 @@ export * from './src/api/rules/validation/validation_errors';
 export * from './src/api/structure';
 export * from './src/api/transformed_value';
 export * from './src/api/types';
-export * from './src/directive/form_field_directive';
-export * from './src/directive/ng_signal_form';
+export * from './src/directive/form_field';
+export * from './src/directive/form_root';
